+++ src/main/java/dooya/see/adapter/webapi/CommentApi.java
@@ -6,8 +6,8 @@ import dooya.see.application.member.required.TokenManager;
 import dooya.see.application.post.provided.CommentFinder;
 import dooya.see.application.post.provided.CommentManager;
 import dooya.see.domain.post.Comment;
-import dooya.see.domain.post.CommentCreateRequest;
-import dooya.see.domain.post.CommentUpdateRequest;
+import dooya.see.domain.post.dto.CommentCreateRequest;
+import dooya.see.domain.post.dto.CommentUpdateRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
