+++ src/main/java/dooya/see/adapter/webapi/PostApi.java
@@ -6,6 +6,9 @@ import dooya.see.application.member.required.TokenManager;
 import dooya.see.application.post.provided.PostFinder;
 import dooya.see.application.post.provided.PostManager;
 import dooya.see.domain.post.*;
+import dooya.see.domain.post.dto.PostCreateRequest;
+import dooya.see.domain.post.dto.PostSearchRequest;
+import dooya.see.domain.post.dto.PostUpdateRequest;
 import dooya.see.domain.post.exception.UnauthorizedPostAccessException;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
