+++ src/main/java/dooya/see/application/post/PostModifyService.java
@@ -5,6 +5,8 @@ import dooya.see.application.post.provided.PostManager;
 import dooya.see.application.post.required.PostLikeRepository;
 import dooya.see.application.post.required.PostRepository;
 import dooya.see.domain.post.*;
+import dooya.see.domain.post.dto.PostCreateRequest;
+import dooya.see.domain.post.dto.PostUpdateRequest;
 import dooya.see.domain.post.exception.UnauthorizedPostAccessException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
