+++ src/main/java/dooya/see/domain/post/Post.java
@@ -1,5 +1,7 @@
 package dooya.see.domain.post;
 
+import dooya.see.domain.post.dto.PostCreateRequest;
+import dooya.see.domain.post.dto.PostUpdateRequest;
 import dooya.see.domain.post.event.*;
 import dooya.see.domain.post.exception.InvalidPostStatusTransitionException;
 import dooya.see.domain.post.exception.UnauthorizedPostAccessException;
