+++ src/test/java/dooya/see/adapter/webapi/CommentApiTest.java
@@ -7,10 +7,10 @@ import dooya.see.application.member.provided.MemberRegister;
 import dooya.see.domain.member.dto.MemberAuthRequest;
 import dooya.see.domain.member.MemberFixture;
 import dooya.see.domain.member.dto.MemberRegisterRequest;
-import dooya.see.domain.post.CommentCreateRequest;
+import dooya.see.domain.post.dto.CommentCreateRequest;
 import dooya.see.domain.post.CommentStatus;
-import dooya.see.domain.post.CommentUpdateRequest;
-import dooya.see.domain.post.PostCreateRequest;
+import dooya.see.domain.post.dto.CommentUpdateRequest;
+import dooya.see.domain.post.dto.PostCreateRequest;
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
