+++ src/test/java/dooya/see/application/post/provided/CommentManagerTest.java
@@ -3,6 +3,8 @@ package dooya.see.application.post.provided;
 import dooya.see.SeeTestConfiguration;
 import dooya.see.application.post.required.CommentRepository;
 import dooya.see.domain.post.*;
+import dooya.see.domain.post.dto.CommentCreateRequest;
+import dooya.see.domain.post.dto.CommentUpdateRequest;
 import dooya.see.domain.post.exception.CommentNotFoundException;
 import dooya.see.domain.post.exception.InvalidCommentStatusException;
 import dooya.see.domain.post.exception.PostNotFoundException;
